lab-materials/04-storage-odf/01-odf-s3/source/requirements.txt
Flask==3.0.3
Werkzeug==3.0.3
Pillow==10.4.0
boto3==1.34.162
gunicorn==22.0.0
